Fix broken indexing with certain embedded systems
The existing indexing code had a tacit assumption that a typical build
of `libclang` would be able to both recognize and work with whatever
target triple is specified by the compiler, in either
compile_commands.json or .ccls files. In cases where libclang was unable
to allocate a target for a given triple, the indexer silently gave up.

This workaround is equivalent to users who have replaced the compiler in
such files post-hoc, or regenerated compile_commands with a host
compiler, where the dependency on the specific target for the purposes
of indexing is effectively removed. If CreateTargetInfo returns nullptr
with the TargetOpts inferred from the compiler name (Most likely because
it couldn't allocate a target for the given triple), the target triple
is overwritten to match the host, and TargetInfo::CreateTargetInfo is
attempted again. Only if that too fails (Something truly bizarre is
happening) will the indexer experience the previous failing behaviour